Moved every visual frames in a own file dynamically loaded according to configuration map
