+++ components/pattern/PatternReactor.js
@@ -9,6 +9,8 @@ const serverConfigRepo = new ServerConfigRepository(
     new DbClient(configEndpoint)
 );
 
+
+
 export default class PatternReactor extends React.Component {
     constructor(props) {
         super(props);
